+++ hypothetical/nonparametric.py
@@ -60,6 +60,7 @@ Wikipedia contributors. (2018, August 22). Wald–Wolfowitz runs test. In Wikipe
 """
 
 from itertools import groupby
+from collections import Counter
 import numpy as np
 import numpy_indexed as npi
 import pandas as pd
@@ -1163,15 +1164,11 @@ class RunsTest(object):
         else:
             self.x = x
 
-        if len(np.unique(self.x)) > 2:
-            raise ValueError('runs test is defined only for runs of binary responses, '
-                             '(True/False, 1/0, Group A/Group B, etc).')
-
         self.runs, self.r = count_runs(self.x)
         self.continuity = continuity
         self.test_summary = self._runs_test()
 
-    def _runs_test(self):
+    def _runs_test(self, n1=None, n2=None):
         r"""
         Primary method for performing the one-sample runs test.
 
@@ -1181,7 +1178,7 @@ class RunsTest(object):
             Dictionary containing relevant test statistics of the one-sample runs test.
 
         """
-        n1, n2 = find_repeats(pd.factorize(self.x)[0]).counts
+        n1, n2 = Counter(self.x).values()
 
         r_range = np.arange(2, self.r + 1)
         evens = r_range[r_range % 2 == 0]
@@ -1774,18 +1771,12 @@ class WaldWolfowitz(object):
 
     """
     def __init__(self, x, y, continuity=True):
-        if not isinstance(x, np.ndarray):
-            self.x = x
-        else:
-            self.x = x
-
-        if not isinstance(y, np.ndarray):
-            self.y = y
-        else:
-            self.y = y
-
+        self.x, self.y = x, y
+        self.n1, self.n2 = len(x), len(y)
+        self.a = np.sort(np.array(self.x + self.y))
         self.continuity = continuity
-        self.r, self.test_summary = self._test()
+        self.r = count_runs(self.a)[1]
+        self.test_summary = self._test()
         self.p_value = self.test_summary['p-value']
         self.probability = self.test_summary['probability']
         self.description = 'Wald-Wolfowitz Runs Test for Two Independent Samples'
@@ -1796,19 +1787,48 @@ class WaldWolfowitz(object):
             pass
 
     def _test(self):
-        x_group = np.repeat('A', len(self.x))
-        y_group = np.repeat('B', len(self.y))
+        r_range = np.arange(2, self.r + 1)
+        evens = r_range[r_range % 2 == 0]
+        odds = r_range[r_range % 2 != 0]
 
-        x = pd.DataFrame({'val': self.x, 'group': x_group})
-        y = pd.DataFrame({'val': self.y, 'group': y_group})
+        p_even = 1 / comb(self.n1 + self.n2, self.n1) * \
+                 np.sum(2 * comb(self.n1 - 1, evens / 2 - 1) *
+                        comb(self.n2 - 1, evens / 2 - 1))
 
-        xy = np.array(x.append(y).sort_values('val'))
+        p_odd = 1 / comb(self.n1 + self.n2, self.n1) * \
+                np.sum(comb(self.n1 - 1, odds - 1) *
+                       comb(self.n2 - 1, odds - 2) +
+                       comb(self.n1 - 1, odds - 2) * comb(self.n2 - 1, odds - 1))
 
-        test = RunsTest(xy[:, 0], continuity=self.continuity)
+        p = p_even + p_odd
 
-        r, test_summary = test.r, test.test_summary
+        if all(np.array([self.n1, self.n2]) <= 20):
+            r_crit_1, r_crit_2 = r_critical_value(self.n1, self.n2)
 
-        return r, test_summary
+            test_summary = {
+                'probability': p,
+                'r critical value 1': r_crit_1,
+                'r critical value 2': r_crit_2
+            }
+            return test_summary
+
+        else:
+            mean = (2 * self.n1 * self.n2) / (self.n1 + self.n2) + 1
+            sd = np.sqrt((2 * self.n1 * self.n2 * (2 * self.n1 * self.n2 - self.n1 - self.n2)) /
+                         ((self.n1 + self.n2) ** 2 * (self.n1 + self.n2 - 1)))
+            z = (np.abs(self.r - mean) - self.continuity * 0.5) / sd
+            p_val = norm.sf(z) * 2
+
+            test_summary = {
+                'probability': p,
+                'mean of runs': mean,
+                'standard deviation of runs': sd,
+                'z-value': z,
+                'p-value': p_val,
+                'continuity': self.continuity
+            }
+
+            return test_summary
 
 
 class WilcoxonTest(object):
@@ -2181,6 +2201,6 @@ def count_runs(x, index=1):
     """
     runs = np.array([sum(1 for _ in r) for _, r in groupby(np.array(x))])
 
-    run_count = len(runs)
+    run_count = np.sum(runs > 1)
 
     return runs, run_count